Create useRouterLoading hook
